+++ storeapp/src/main/java/com/localweb/storeapp/repository/ClientRepository.java
@@ -9,5 +9,5 @@ import java.util.List;
 
 @Repository
 public interface ClientRepository extends JpaRepository<Client, Integer> {
-    Client findClientById(int id);
+    Client findClientById(long id);
 }
